+++ src/lib/waku_message/index.ts
@@ -38,7 +38,6 @@ export interface Options {
   sigPrivKey?: Uint8Array;
 }
 
-// TODO: Use this in Options
 export interface DecryptionParams {
   key: Uint8Array;
   method?: DecryptionMethod;
@@ -270,8 +269,7 @@ export class WakuMessage {
   }
 
   get version(): number {
-    // TODO: absent value should be replaced by default
-    // value of the type by the protobuf decoder
+    // https://github.com/status-im/js-waku/issues/921
     return this.proto.version ?? 0;
   }