+++ src/lib/waku_store/index.node.spec.ts
@@ -228,10 +228,8 @@ describe("Waku Store", () => {
     nwaku = new Nwaku(makeLogFileName(this));
     await nwaku.start({ persistMessages: true, store: true, lightpush: true });
 
-    const encryptedAsymmetricMessageText =
-      "This message is encrypted for me using asymmetric";
-    const encryptedSymmetricMessageText =
-      "This message is encrypted for me using symmetric encryption";
+    const encryptedAsymmetricMessageText = "asymmetric encryption";
+    const encryptedSymmetricMessageText = "symmetric encryption";
     const clearMessageText =
       "This is a clear text message for everyone to read";
     const otherEncMessageText =
@@ -304,14 +302,12 @@ describe("Waku Store", () => {
 
     dbg("Retrieve messages from store");
     const messages = await waku2.store.queryHistory([], {
-      decryptionKeys: [privateKey],
+      decryptionParams: [{ key: privateKey }],
     });
 
-    expect(messages?.length).eq(3);
-    if (!messages) throw "Length was tested";
-    expect(messages[0].payloadAsUtf8).to.eq(clearMessageText);
-    expect(messages[1].payloadAsUtf8).to.eq(encryptedSymmetricMessageText);
-    expect(messages[2].payloadAsUtf8).to.eq(encryptedAsymmetricMessageText);
+    expect(messages[0]?.payloadAsUtf8).to.eq(clearMessageText);
+    expect(messages[1]?.payloadAsUtf8).to.eq(encryptedSymmetricMessageText);
+    expect(messages[2]?.payloadAsUtf8).to.eq(encryptedAsymmetricMessageText);
 
     !!waku1 && waku1.stop().catch((e) => console.log("Waku failed to stop", e));
     !!waku2 && waku2.stop().catch((e) => console.log("Waku failed to stop", e));
@@ -411,7 +407,7 @@ describe("Waku Store", () => {
 
     dbg("Retrieve messages from store");
     const messages = await waku2.store.queryHistory([], {
-      decryptionKeys: [privateKey],
+      decryptionParams: [{ key: privateKey }],
     });
 
     expect(messages?.length).eq(3);